requirements.txt
torch
torchvision
Pillow>=9.0.0
tqdm
numpy>=1.21.0